refactoring chips and fixing input
